disable product info & list

+++ service/internal/core/usecase/chart/member_chart.go
@@ -5,6 +5,7 @@ import (
 
 	"golang.org/x/xerrors"
 
+	"simon/mall/service/internal/constant"
 	"simon/mall/service/internal/errs"
 	"simon/mall/service/internal/model/bo"
 	"simon/mall/service/internal/model/po"
@@ -49,12 +50,23 @@ func (uc *memberChartUseCase) GetMemberChart(ctx context.Context) ([]*bo.MemberC
 	result := make([]*bo.MemberChart, len(charts))
 	// use chart & product combination result
 	for i := 0; i < len(charts); i++ {
+		// todo 檢查 chart對應不上product
 		result[i] = &bo.MemberChart{
 			Id:       charts[i].Id,
-			Name:     products[charts[i].ProductId].Name,
-			Amount:   products[charts[i].ProductId].Amount,
 			Quantity: charts[i].Quantity,
-			Image:    products[charts[i].ProductId].Image,
+		}
+
+		// 預期外的購物車產品
+		if val, ok := products[charts[i].ProductId]; !ok {
+			result[i].Name = constant.Unknown_Product
+			result[i].Amount = 0
+			result[i].Image = ""
+			result[i].Inventory = 0
+		} else {
+			result[i].Name = val.Name
+			result[i].Amount = val.Amount
+			result[i].Image = val.Image
+			result[i].Inventory = val.Inventory
 		}
 	}
 
@@ -69,6 +81,10 @@ func (uc *memberChartUseCase) UpdateMemberChart(ctx context.Context, cond *bo.Me
 		return errs.MemberTokenError
 	}
 
+	if err := uc.validateUpdate(ctx, cond); err != nil {
+		return xerrors.Errorf("memberChartUseCase.UpdateMemberChart -> validateUpdate : %w", err)
+	}
+
 	db := uc.in.DB.Session()
 	if err := uc.in.MemberChartRepo.Update(ctx, db, &po.MemberChartUpdate{
 		Id:       cond.Id,
@@ -76,6 +92,18 @@ func (uc *memberChartUseCase) UpdateMemberChart(ctx context.Context, cond *bo.Me
 		Quantity: cond.Quantity,
 	}); err != nil {
 		return xerrors.Errorf("memberChartUseCase.UpdateMemberChart -> ProductCommon.GetProduct : %w", err)
+	}
+
+	return nil
+}
+
+func (uc *memberChartUseCase) validateUpdate(ctx context.Context, cond *bo.MemberChartUpdateCond) interface{} {
+	if len(cond.Id) == 0 {
+		return xerrors.Errorf("memberChartUseCase.UpdateMemberChart -> cond.Id == 0: %w", errs.RequestParamInvalid)
+	}
+
+	if cond.Quantity <= 0 {
+		return xerrors.Errorf("memberChartUseCase.UpdateMemberChart -> cond.Quantity <= 0: %w", errs.RequestParamInvalid)
 	}
 
 	return nil
@@ -137,9 +165,15 @@ func (uc *memberChartUseCase) validateCreate(ctx context.Context, cond *bo.Membe
 func (uc *memberChartUseCase) DeleteMemberChart(ctx context.Context, cond *bo.MemberChartDelCond) error {
 	defer timelogger.LogTime(ctx)()
 
+	memberInfo, ok := ctxs.GetSession(ctx)
+	if !ok {
+		return errs.MemberTokenError
+	}
+
 	db := uc.in.DB.Session()
 	if err := uc.in.MemberChartRepo.Delete(ctx, db, &po.MemberChartDel{
-		Id: cond.Id,
+		Id:       cond.Id,
+		MemberId: memberInfo.Id,
 	}); err != nil {
 		return xerrors.Errorf("memberChartUseCase.DeleteMemberChart -> MemberChartRepo.Delete : %w", err)
 	}
